feat(search redux): created redux for search products
