+++ frontend/src/redux/products/operations.js
@@ -65,19 +65,3 @@ export const getProductsId = createAsyncThunk(
     },
   }
 );
-
-export const searchProducts = createAsyncThunk(
-  'product/searchProducts',
-  async (query, thunkAPI) => {
-    try {
-      const res = await axios.get(`${baseApiUrl}/search/`, {
-        params: { query },
-      });
-
-      return res.data;
-    } catch (e) {
-      console.error('Error searching products:', e);
-      return thunkAPI.rejectWithValue(e.message);
-    }
-  }
-);
