+++ frontend/src/redux/products/slice.js
@@ -1,5 +1,5 @@
 import { createSlice } from '@reduxjs/toolkit';
-import { getProducts, getProductsId, searchProducts } from './operations';
+import { getProducts, getProductsId } from './operations';
 
 const handlePending = (state) => {
   state.loading = true;
@@ -21,32 +21,6 @@ const productSlice = createSlice({
     error: false,
     createSuccess: false,
     createdProduct: null,
-
-    //SEARCH POPUP
-    query: '',
-    searchResults: [],
-    categories: [],
-    searchHistory: [],
-  },
-  reducers: {
-    clearSearch: (state) => {
-      state.query = '';
-      state.searchResults = [];
-      state.categories = [];
-    },
-    clearSearchHistory: (state) => {
-      state.searchHistory = [];
-    },
-    addSearchHistory: (state, action) => {
-      const newItem = action.payload;
-      if (!newItem?.id || !newItem?.title?.trim()) return;
-
-      const filtered = state.searchHistory.filter(
-        (item) => item.id !== newItem.id
-      );
-
-      state.searchHistory = [newItem, ...filtered].slice(0, 3);
-    },
   },
   extraReducers: (builder) => {
     builder
@@ -65,21 +39,8 @@ const productSlice = createSlice({
         state.loading = false;
         state.error = false;
       })
-      .addCase(getProductsId.rejected, handleRejected)
-
-      .addCase(searchProducts.pending, handlePending)
-      .addCase(searchProducts.fulfilled, (state, action) => {
-        state.query = action.payload.query;
-        state.searchResults = action.payload.products;
-        state.categories = action.payload.categories;
-        state.loading = false;
-        state.error = false;
-      })
-      .addCase(searchProducts.rejected, handleRejected);
+      .addCase(getProductsId.rejected, handleRejected);
   },
 });
-
-export const { clearSearch, addSearchHistory, clearSearchHistory } =
-  productSlice.actions;
 
 export const productReducer = productSlice.reducer;
