+++ frontend/src/redux/searchProducts/selectors.js
@@ -1,0 +1,7 @@
+export const selectSearchResults = (state) => state.search.searchResults;
+
+export const selectSearchQuery = (state) => state.search.query;
+
+export const selectSearchCategories = (state) => state.search.categories;
+
+export const selectSearchHistory = (state) => state.search.searchHistory;
